Initializing repo with files for rmlands writeup
Have been working on this a while but not yet set up a repository. Planning to make some changes today (12/01/13) and wanted to have a backup in place for monitoring changes down the line.

At this point everything from the lab notebook up through about the end of October is in one of these associated tex files.

	new file:   1_InputLayerPrep.tex
	new file:   2_DisturbanceSuccessionTransitionDevelopment.tex
	new file:   3_Procedures.tex
	new file:   4_RMLandsParameterDevelopment.tex
	new file:   5_ModelCalibration.tex
	new file:   PDSIPointMap.pdf
	new file:   PDSIpic.png
	new file:   WeatherStationMap.pdf
	new file:   master.tex